lexiops-copilot/motia/requirements.txt
openai>=1.54.4
aiohttp>=3.8.0
python-dotenv>=1.0.0